refactor: share stylesheet between components to avoid duplication
