--- src/views/instances/new-dhis2/new-dhis2-instance-form.module.css
@@ -1,21 +1,0 @@
-.legend {
-    font-size: 20px;
-    font-weight: 500;
-    line-height: 23px;
-    padding-bottom: var(--spacers-dp24);
-}
-.fieldset {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: var(--spacers-dp16);
-    border: none;
-    padding: 0;
-    margin: 0;
-}
-.hr {
-    border: none;
-    border-top: 1px solid var(--colors-grey400);
-    width: 100%;
-    margin: var(--spacers-dp24) 0;
-}
