+++ src/views/instances/new-dhis2/new-dhis2-instance-form.tsx
@@ -1,12 +1,13 @@
 import { Button, ButtonStrip } from '@dhis2/ui'
+import cx from 'classnames'
 import { AnyObject } from 'final-form'
 import { DescriptionTextarea } from './fields/description-textarea'
 import { GroupSelect } from './fields/group-select'
 import { NameInput } from './fields/name-input'
 import { PublicCheckbox } from './fields/public-checkbox'
 import { TtlSelect } from './fields/ttl-select'
-import styles from './new-dhis2-instance-form.module.css'
 import { ParameterFieldset } from './parameter-fieldset'
+import styles from './styles.module.css'
 
 type NewDhis2InstanceFormProps = {
     handleCancel: () => void
@@ -18,7 +19,7 @@ type NewDhis2InstanceFormProps = {
 
 export const NewDhis2InstanceForm = ({ handleCancel, handleSubmit, submitting, pristine, invalid }: NewDhis2InstanceFormProps) => (
     <form onSubmit={handleSubmit}>
-        <fieldset className={styles.fieldset}>
+        <fieldset className={cx(styles.fieldset, styles.main)}>
             <legend className={styles.legend}>Basic information</legend>
             <NameInput />
             <DescriptionTextarea />
