--- src/views/instances/new-dhis2/new-dhis2-instance.module.css
@@ -1,10 +1,0 @@
-.container {
-    display: block !important;
-    height: auto !important;
-    padding: var(--spacers-dp16);
-}
-.header {
-    font-size: 28px;
-    font-weight: 400;
-    line-height: 40px;
-}
