+++ src/views/instances/new-dhis2/styles.module.css
@@ -1,29 +1,57 @@
+.container {
+    display: block !important;
+    height: auto !important;
+    padding: var(--spacers-dp16);
+}
+
+.header {
+    font-size: 28px;
+    font-weight: 400;
+    line-height: 40px;
+}
+
 .legend {
     font-size: 20px;
     font-weight: 500;
     line-height: 23px;
     padding-bottom: var(--spacers-dp24);
 }
+
 .fieldset {
     border: none;
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    gap: 12px;
+    gap: var(--spacers-dp16);
     padding: 0;
     margin: 0;
 }
-.fieldset.primary {
+
+.fieldset.main {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    padding: 0;
+    margin: 0;
+}
+
+.fieldset.parameters {
+    display: grid;
+}
+
+.fieldset.parameters.primary {
     grid-template-columns: 320px 320px;
 }
-.fieldset.secondary {
+
+.fieldset.parameters.secondary {
+    grid-template-columns: repeat(3, 1fr);
     padding-top: var(--spacers-dp16);
 }
+
 .hr {
     border: none;
     border-top: 1px solid var(--colors-grey400);
     width: 100%;
     margin: var(--spacers-dp24) 0;
 }
+
 .summary {
     color: var(--colors-blue600);
     margin: var(--spacers-dp24) 0 0 0;
